Brainstorm pseudocode
This commit adds pseudocode to start building the code from. The first
line I want to start with is 'array.length > 1' to see if a given array
or subarray will need to be separated further. If the length is < 1 then
the merge/sort process can start in the else block.