Remove 'sorted' parameter, add merge arrays
This commit remove the 'sorted' parameter from the function definition.
Instead three empty arrays are added after the base/recursive case if..
else block. These arrays are meant to serve as a way of containing
one element each to be compared for sorting.

The sorting section hasn't been coded yet because merge_b is not
displaying its designated element.