Add merge.rb file
This commit adds a file called merge.rb. This file was created separate
from the main rb file because I wanted to try working out a merge
solution in a separate environment.

This code is implementing variables to select elements in the
two starting arrays (beginning with the first) and progresses after each
comparison.

The current issue is with line 12 'if (a[i] < b[j])'. An undefined
method error is raised saying that there is a nil in this comparison.
I've tried writing the line as 'if (a[i] < b[0])' and the array, 'c' is
printed as follows:

  [nil, 2, 4, 5, 6, nil, nil]

At first I thought there might be an issue with the while loop's scope
but a while loop's scope is the same as the code containing it.