Change while loop condition
This commit changed the while loop condition that handles the comparison
between two elements in arrays 'a' and 'b'. In the previous version, I
kept getting an error due to a comparison with a nil value. After
changing "while (i <= x && j <= y) do" to "until (i == x || j == y) do"
the errors weren't being raised.

There is also an if..else block added after the previous block to handle
elements that are left over once one array reaches its end. This block
isn't correct but I wrote it in to see how it carries out the rest of
the function. Array 'c' prints out:

  [nil, 2, 4, 5, nil, 6]

There is still more corrections to be made but getting the correct
number of elements seemed like a good sign.