Undo changes to loop condition
This commit reverts the 'until' loop back to a 'while' loop, as it was
initially. Incrementing variables, 'i', 'j', and 'k' were written as
separate lines because I wasn't sure if they were causing issues with
indexes as the array elements were being compared. The print command in
line 21 shows that array 'c' is taking in the elements in correct order
(given that 'a' and 'b' are ordered lists).

The one issue is with the final element of array 'b', which is 4. Since
this is the last value, it means that the other array has reached its
end but the 4 isn't being copied over to array 'c'. There is an error
stating that a comparison between an Integer and nil doesn't work.