While loop, '<=' changed to '<'
This commit changes the comparison operator from '<=' to '<'. Since the
length of an array is larger than the last index of an array by 1, I
wanted a way to exit the loop when those numbers were the same. This
would mean that one of the arrays would be at its end and the code can
move to the next block.

In this version I commented out the block handling the remaining element(s)
and wrote another one that checks if the next value is nil. If one array
has a nil value, the other array's value gets sent to array 'c'. I still
need to code a way to select the rest of the values in an array rather
than just where the index is at.