Insert remaining elements into array 'c'
This commit changes lines 32 - 35, the section that handles the remaining
elements of an array when the other array reaches a nil value.

Previously, when a nil value comes up, only one element of the other
array was copied into array 'c'. Now the elements from wherever the
index is to the end of that array is copied over. However, the final
result had a nested array inside, hence the #flatten in line 38.

Currently the code works for a pair of arrays with three elements (given
that elements are in order) but it will not work for arrays with only
one element. This will have to be fixed since the idea behind the
recursive function is to get the smallest pieces to sort and merge.