Require merge.rb
This commit incorporates the merge.rb file into the original file I was
working on. This file now only has to determine if an array has less
than two elements and call the merge function if necessary.

The merge sort is unsuccessful with arrays that are larger than two
elements each.