Add block to check for nil
This commit adds an if block to check if an incoming argument to the
merge function is nil. This was put in to prevent errors from comparing
nil values in merge.rb.

However, in merge_sort.rb, the result is two separate arrays that have
elements in order, not one complete array with elements in order.