Add result array default argument
This commit adds an empty array, 'result', as a default parameter at the
start of the method. This is where the results of the merge function
get stored in after each 'a' and 'b' halves are created.

Within the else block, a small if block checks to see if the resulting
'c' array has any truthy elements to pass to the 'result' array.

After adding these lines, the 'result' array shows that it contains
nested arrays [1,3] and [2,4]. However, an error is thrown in the
merge.rb file right after.