Call merge with result array
This commit adds another call to the merge function with the 'result'
array. The 'a' and 'b' arrays are reassigned to halves of this array
before the function is called.

Though the merge_sort function works with a starting array of four
elements, it has trouble with arrays of different numbers. Results often
look nearly complete with a couple elements out of order. In the case of
a five element starting array, the result is four elements.