Add merge_sort_final
This commit adds merge_sort_final.rb, which is a combination of merge.rb,
merge_sort.rb along with a lot of changes and reduction of code.

Merge_sort_final.rb was my way of trying to recreate this function after
struggling with finding the solution to this assignment. I ended up
looking at a few others' solutions and noticed that my code was
unnecessarily long and cluttered. I also discovered that some of my
errors were due to the effects of using '<<' vs. '+' to combine arrays.

Another part of my code that might have caused a lot of issues and
clutter was using indeces to compare elements. I instead resorted to the
shift method which simultaneously returns the first element and removes
it. This way, I didn't have to worry about incrementing indeces for
three different arrays.

Overall, I had a good grasp of the understanding of how this function
was supposed to work. If I had been more careful and selective with my
array methods and looked into what was being returned I believe I may
have had a more complete solution, regardless of the length of my code.